# speed up plugin update