@Rechi 2019-05-18
https://jenkins.kodi.tv/job/BuildMulti-PR/12186/console
> ERROR: This PR contains 2 merge commit(s)